requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data Processing and Analysis
pandas==2.1.3
openpyxl==3.1.2
numpy==1.24.3

# Math and Algorithms
scipy==1.11.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and Testing
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional (for production)
gunicorn==21.2.0